requirements.txt
alembic==1.4.3
aniso8601==8.0.0
apispec==4.0.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
flask-apispec==0.11.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kanpai==0.1.15
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.19
visitor==0.1.3
webargs==6.1.1
Werkzeug==1.0.1
WTForms==2.3.3
